new Order

+++ WebApplication1/Data/Order.cs
@@ -17,7 +17,7 @@ namespace OrderRestaurant.Data
         public DateTime? CreationTime { get; set; }
         public DateTime? ReceivingTime { get; set; }
         public DateTime? PaymentTime { get; set; }
-        public double? Pay {  get; set; }
+        public decimal? Pay {  get; set; }
         public string? Note { get; set; }
         
         public int StatusId { get; set; }
